Add files via upload
The main objective of this deep learning project is to develop a predictive model capable of accurately classifying images of animals as either cats or dogs. Leveraging deep learning algorithms, we aim to create a robust and efficient system that can automatically recognize and categorize these common animal species based on visual characteristics.